tools/docs-translator/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docs-translator"
version = "1.0.0"
description = "AI-powered documentation translator with Markdown structure preservation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "That Le", email = "team@agrios.com"}
]
keywords = ["translation", "markdown", "documentation", "ai", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Documentation",
    "Topic :: Text Processing :: Markup :: Markdown",
]

dependencies = [
    "markdown-it-py>=3.0.0",
    "google-generativeai>=0.3.0",
    "openai>=1.0.0",
    "anthropic>=0.8.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "keyring>=24.0.0",
    "inquirer>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "mypy>=1.7.0",
    "ruff>=0.1.0",
]

[project.scripts]
docs-translator = "docs_translator.cli.main:cli"

[project.urls]
Homepage = "https://github.com/agrios/docs-translator"
Documentation = "https://github.com/agrios/docs-translator#readme"
Repository = "https://github.com/agrios/docs-translator.git"

[tool.setuptools.packages.find]
where = ["."]
include = ["docs_translator*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
